stats page modularized
